fix/disable some options when looking at past activity and add tip
Disables options that should not be possible when looking at past activity entries, and adds tool tip which says "Select latest activity to enable." No testing has been done yet as this is a UI change, and I am not prepared for a pull request until looking it over more fully. For now will be switching to work on another issue, will be back on this one soon.